Add stamped data to the callback params.
The problem is that the stamped data belongs with
the result string. With multiple parallel 
renders happening, we cannot guarantee 
render order, so attaching the slot data to the
template function will not always work.

Also, add a finished flag to the render data,
so we know when the render is complete.