Conditional queries for include,>,extends,block:
For these tags, we might want to pass in a string
literal, or lookup something on the context. If
we prefix our key with a #, it will use context
lookup, otherwise, it will be a string literal.

Tags that don't make sense with a string literal
(#,if,stamp) do not require the # prefix.

So:
.# content blog/first-post
  .< {_type}
./#

Would render a partial named with the content
object's type; (eg: 'post')

This lets us do some dynamic rendering that is
not possible with just string literals.